+++ menu/src/render/draw_util.hpp
@@ -24,7 +24,7 @@ namespace base::render::util {
   ImVec2 CalcTextSize(const ImFont* font, float font_size, const std::string& text, float wrap_width = 0.0f);
 
   // TODO: Make a non monospace version of this.
-  std::uint32_t WordWrap(float font_size, std::string& str, float max_x, std::uint32_t max_lines);
+  std::uint32_t WordWrap(float font_size, std::string& str, float max_x, std::size_t max_lines);
 
   ImDrawList* GetDrawList();
 }
